+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 * [NOTICE] The interactive links (like in `PCB`) are now not possible in EVA construction mode.
 * [Change] Better detect if any of the peers in the KAS connection got destroyed for any reason. The link gets properly broken in this case.
 * [Enhancement] Allow attaching to the winches surface to let reinforcing them with struts.
-* [Enhancement] Scale RTS dialog to match the game's GUI scale.
 * [Enhancement] Don't show resources that cannot be transferred in the RTS GUI.
+* [Enhancement #248] Add ability to scale Transfer GUI.
 * [Fix #302] GUI does not respect hide/show function.
 * [Fix #306] Logs spam from the parts dropped on the ground.
 * [Fix #307] Interactive attach mode conflicts with construction mode.